removed all existing algorithm code and replaced with provider
